Move out Settings editor indicator label impl
